+++ ui/src/components/session/sessionDialog/connections/SessionConnectionDockerContainer.tsx
@@ -4,7 +4,7 @@ import React, { useEffect, useState } from 'react'
 import { DockerDesktopClient } from '@docker/extension-api-client-types/dist/v1'
 import { ContainerInfo } from '../../../../types/docker/extension'
 import ContainerSelect from '../../../connectbar/ContainerSelect'
-import { serializeConnectionData } from '../SessionDialog'
+import { serializeConnectionData } from '../../forms/SessionDataForm'
 
 
 interface DockerContainerProps {