+++ ui/src/components/session/sessionDialog/connections/SessionConnectionRemoteHost.tsx
@@ -1,6 +1,6 @@
 import { FormGroup, Stack, TextField } from '@mui/material'
 import React, { useEffect, useState } from 'react'
-import { serializeConnectionData } from '../SessionDialog'
+import { serializeConnectionData } from '../../forms/SessionDataForm'
 
 
 interface RemoteHostProps {
@@ -55,7 +55,6 @@ export default function SessionConnectionRemoteHost ({ connectionData, setSubmit
           label="Remote Host IP/NAME"
           value={remoteHost}
           onChange={event => setRemoteHost(event.target.value)}
-          autoFocus
           required
         />
         <TextField